fixed the minimax
It actually finds the best move now instead of always going for the last one checked. It did that because the alpha and beta where reset to -oo and +oo when it got back to the root node. It now saves the best score and move for each branch and finds the best one once it has searched them all.